A princípio consegui corrigir e finalizar o que estava faltando, porém quando rodei deu um erro de bean. Por isso resolvi commitar o que fiz até agora e então vou mexer pra tentar resolver esse erro sem danificar o que está funcionando.

+++ src/main/java/com/grupo5/monolitoclean/controllers/aplicativos/AplicativoController.java
@@ -27,7 +27,7 @@ public class AplicativoController {
     }
 
     @PostMapping("/atualizacusto/{idAplicativo}")
-    public AplicativoDTO atualizaCusto(@PathVariable long idAplicativo, @RequestBody double custoMensal){
+    public AplicativoDTO atualizaCusto(@PathVariable long idAplicativo, @RequestBody float custoMensal){
         return atualizaCusto.run(idAplicativo, custoMensal);
     }
 }
